pyproject.toml
[project]
name = "amhe"
version = "0.1.0"
description = "des algorithm with surrogate model to limit the evaluation function runs"
authors = [
    {name = "Filip",email = "fbudzynski56@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.2.6,<3.0.0)",
    "deap (>=1.4.3,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "cocopp (>=2.7.3,<3.0.0)",
    "coco-experiment (>=2.8.1,<3.0.0)"
]

packages = [{ include = "amhe", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
